RSB-9 refactor: structure for lib and styles
