+++ components.json
@@ -5,7 +5,7 @@
   "tsx": true,
   "tailwind": {
     "config": "tailwind.config.ts",
-    "css": "src/app/globals.css",
+    "css": "src/styles/globals.css",
     "baseColor": "neutral",
     "cssVariables": true,
     "prefix": ""
@@ -13,7 +13,8 @@
   "aliases": {
     "components": "@/components",
     "ui": "@/components",
-    "utils": "@/lib/utils",
-    "lib": "@/lib"
+    "utils": "@/utils",
+    "lib": "@/lib",
+    "styles": "@/styles"
   }
 }
