+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import { config } from "@fortawesome/fontawesome-svg-core";
-import "./globals.css";
+import "@/styles/globals.css";
 import "@fortawesome/fontawesome-svg-core/styles.css";
 config.autoAddCss = false;
